Jenkins/requirements.txt
docopt-dispatch==0.0.2
xmltodict==0.11.0
docopt-dispatch==0.0.2
PyYAML==3.12
ruamel.yaml==0.15.74
